GUA-426: Rename entrypoint file
We're using the lambda name as the entrypoint file name to be
consistent.
